Add Assessment and relevant data field classes to seedu.address.model.assessment package
An Assessment contains AssessmentName, Weightage and FullMark fields. It does not contain the student's results. The reason for this is because a non-Tutorial-specific Assessment class is helpful if the team chooses to implement commands like `list assessments` or `edit assessment`.

Each Assessment is associated with an AssessmentResults (present in a later commit).