Add UniqueAssessmentList and related exceptions to seedu.address.model.assessment package
UniqueAssessmentList is a list containing all the Assessment objects in the address book. It enforces uniqueness among the elements using the AssessmentName as the identifier.

This class is useful for listing assessments or editing an assessment.

Its backing list is an ObservableList which has the benefit of allowing listeners to track changes to the list (so that the UI can update the list of assessments automatically when there is a change in the UniqueAssessmentList).

See UniquePersonList in the model.person package and PersonListPanel in the ui package for reference.