Add AssessmentResultsList, AssessmentResults and StudentResult classes to seedu.address.model.assessment package
StudenResult represents the result of a single student for an assessment and contains an NusNetId and Score.
AssessmentResults contains a list of StudentResults and has an AssessmentName (is associated with an Assessment).
AssessmentResultsList contains a list of AssessmentResults and has a TutorialName (inside a Tutorial).

An AssessmentResultsList is associated with a Tutorial while the UniqueAssessmentList is associated with the ModelManager or AddresBook.

AssessmentResults and AssessmentResultsList both contain ObservableLists for the same reason as in the previous commit.

Note: the AssessmentNotFoundException is meant for the previous commit but DuplicateStudentResultException is added in the previous commit by accident.