Add a field called allStudents, which contains all students in the address book to ModelManager class
allStudents is declared as a FilteredList<Person>, which takes in the address book's persons list and a predicate to filter the students.

allStudents is automatically updated when the address book's person list is updated (this has been manually tested on a smaller scale) so there is no need to manually update allStudents.

More changes need to be done to ModelManager and/or AddressBook to include new data classes.