Add a tutorial list and assessment list to AddressBook class to integrate new data classes into model
AddressBook wraps all the data at the address-book level, including the persons, tutorials and assessments. It provides methods for updating the lists of persons, tutorials and assessments which are used by the ModelManager to update the model components after a command is executed.

Todo: add methods for updating the tutorial list and assessment list in the AddressBook and ModelManager classes (update the Model interface as well).