Change the logic and model components to support adding an assessment to the address book
Add 2 new classes to the logic component: AddAssessmentCommand and AddAssessmentCommandParser to support the execution and parsing of add assessment commands respectively. Update AddressBookParser to handle add assessment commands. Add the prefixes of add assessment command to CliSyntax class.

Update ModelManager, Assessment, UniqueTutorialList, Tutorial to support the adding of an assessment to the addressbook. Apart from adding an Assessment object to the UniqueAssessmentList in the AddressBook, the AssessmentResultsList of every Tutorial in the UniqueTutorialList needs to be updated with a new AssessmentResults object.